행성 설정 페이지 경로 수정
